Rename module_utils/vmware_guest.py
