+++ plugins/module_utils/vm_device_helper.py
